+++ apps/web/res/css/_components.pcss
@@ -72,7 +72,6 @@
 @import "./structures/_PictureInPictureDragger.pcss";
 @import "./structures/_QuickSettingsButton.pcss";
 @import "./structures/_RightPanel.pcss";
-@import "./structures/_RoomSearch.pcss";
 @import "./structures/_RoomView.pcss";
 @import "./structures/_SearchBox.pcss";
 @import "./structures/_SpaceHierarchy.pcss";
@@ -259,8 +258,6 @@
 @import "./views/rooms/_IRCLayout.pcss";
 @import "./views/rooms/_InvitedIconView.pcss";
 @import "./views/rooms/_JumpToBottomButton.pcss";
-@import "./views/rooms/_LegacyRoomList.pcss";
-@import "./views/rooms/_LegacyRoomListHeader.pcss";
 @import "./views/rooms/_LiveContentSummary.pcss";
 @import "./views/rooms/_MemberListHeaderView.pcss";
 @import "./views/rooms/_MemberListView.pcss";
@@ -277,15 +274,12 @@
 @import "./views/rooms/_ReadReceiptGroup.pcss";
 @import "./views/rooms/_ReplyPreview.pcss";
 @import "./views/rooms/_ReplyTile.pcss";
-@import "./views/rooms/_RoomBreadcrumbs.pcss";
 @import "./views/rooms/_RoomHeader.pcss";
 @import "./views/rooms/_RoomInfoLine.pcss";
 @import "./views/rooms/_RoomKnocksBar.pcss";
 @import "./views/rooms/_RoomPreviewBar.pcss";
 @import "./views/rooms/_RoomPreviewCard.pcss";
 @import "./views/rooms/_RoomSearchAuxPanel.pcss";
-@import "./views/rooms/_RoomSublist.pcss";
-@import "./views/rooms/_RoomTile.pcss";
 @import "./views/rooms/_RoomUpgradeWarningBar.pcss";
 @import "./views/rooms/_SendMessageComposer.pcss";
 @import "./views/rooms/_Stickers.pcss";
