--- apps/web/res/css/views/rooms/_LegacyRoomList.pcss
@@ -1,11 +1,0 @@
-/*
-Copyright 2024 New Vector Ltd.
-Copyright 2020 The Matrix.org Foundation C.I.C.
-
-SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only OR LicenseRef-Element-Commercial
-Please see LICENSE files in the repository root for full details.
-*/
-
-.mx_LegacyRoomList {
-    padding-right: 7px; /* width of the scrollbar, to line things up */
-}
